+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.1.4(빌드 안됨)
+
+* Editor를 GUI로 구현하기 위해 루트 Cargo와 editor/Cargo에 의존성을 추가하였습니다.
+* editor/src/main을 GUI버전으로 변경하였습니다.
+* runner에 키보드 후킹을 추가했습니다.
+* shared의 config에 clone을 추가하였습니다.
+
+---
+
 ## 1.1.0
 
 * Editor로 매크로 등록, 삭제, 추가를 할 수 있습니다.